requirements.txt
python-telegram-bot == 11.1.0 # Python API for Telegram Bot

babel == 2.6.0 # Localization library with gettext support

sqlalchemy == 1.2.11
alembic == 1.0.0
psycopg2 == 2.7.5
